% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-server.R
\name{areNamesNotMatching}
\alias{areNamesNotMatching}
\title{Are Names Not Matching}
\usage{
areNamesNotMatching(
  values,
  entryName,
  newNames,
  isNamesFun = isTargetNames,
  n = NULL
)
}
\arguments{
\item{values}{(list) with all data and model options}

\item{entryName}{(character) name of values element}

\item{newNames}{(reactive) character vector of newNames}

\item{isNamesFun}{(function) function that checks for the correct level in the list hierarchy}

\item{n}{(numeric) depth of list to look for names
of values}
}
\description{
Check if names of entries are matching new newNames
}
