% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkDataDimensions}
\alias{checkDataDimensions}
\title{Check if data dimensions are correct}
\usage{
checkDataDimensions(data, valueNames, modelOptions, argName = NULL)
}
\arguments{
\item{data}{list: input data}

\item{valueNames}{list: contains names for the \code{targets},
\code{fractions}, and \code{sources}}

\item{modelOptions}{list: contains elements \code{burnin} and \code{iterations}}

\item{argName}{character: name of argument displayed in error message}
}
\description{
Check if data dimensions are correct
}
