% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkDataNameConsistency}
\alias{checkDataNameConsistency}
\title{Check if data names are correct}
\usage{
checkDataNameConsistency(data, modelOptions)
}
\arguments{
\item{data}{list: input data}

\item{modelOptions}{list: contains elements \code{burnin} and \code{iterations}}
}
\description{
Check if data names are correct
}
