% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkDim}
\alias{checkDim}
\title{Check if dimensions are correct}
\usage{
checkDim(x, dimExpected, argName = NULL)
}
\arguments{
\item{x}{any object}

\item{dimExpected}{numeric: vector of expected dimensions}

\item{argName}{character: name of argument displayed in error message}
}
\description{
Check if dimensions are correct
}
