% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkIndividualNames}
\alias{checkIndividualNames}
\title{Check individual names}
\usage{
checkIndividualNames(data)
}
\arguments{
\item{data}{list: input data}
}
\description{
Check if indibividual names are given and assign them if missing
}
