% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkLength}
\alias{checkLength}
\title{Check if length is correct}
\usage{
checkLength(x, lengthExpected, argName = NULL)
}
\arguments{
\item{x}{any object}

\item{lengthExpected}{numeric or integer of length 1: expected length}

\item{argName}{character: name of argument displayed in error message}
}
\description{
Check if length is correct
}
