% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkModelOptions}
\alias{checkModelOptions}
\title{Check model options}
\usage{
checkModelOptions(modelOptions)
}
\arguments{
\item{modelOptions}{list: contains elements \code{burnin} and \code{iterations}}
}
\description{
Check model options and its elements for correct class and names
}
