% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkNames}
\alias{checkNames}
\title{Check if names are correct}
\usage{
checkNames(x, namesExpected, argName = NULL)
}
\arguments{
\item{x}{any object}

\item{namesExpected}{character: vector of expected names}

\item{argName}{character: name of argument displayed in error message}
}
\description{
Check if names are correct
}
