% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkPriors}
\alias{checkPriors}
\title{Check priors}
\usage{
checkPriors(priors)
}
\arguments{
\item{priors}{list of characters with prior expressions}
}
\description{
Check priors for correct class
}
