% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkUserEstimates}
\alias{checkUserEstimates}
\title{Check user estimates}
\usage{
checkUserEstimates(userEstimates)
}
\arguments{
\item{userEstimates}{list of characters}
}
\description{
Check user estimates for correct class
}
