% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{checkValueNames}
\alias{checkValueNames}
\title{Check value names}
\usage{
checkValueNames(valueNames)
}
\arguments{
\item{valueNames}{list: contains names for the \code{targets},
\code{fractions}, and \code{sources}}
}
\description{
Check value names and its elements for correct class, length and
names
}
