% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{componentsServer}
\alias{componentsServer}
\title{Components Server}
\usage{
componentsServer(id, values, events)
}
\arguments{
\item{id}{id of module}

\item{values}{values}

\item{events}{events}

\item{hideTargetFilter}{hideTargetFilter}

\item{sourceObsvnFilterChoices}{sourceObsvnFilterChoices}

\item{sourceObsvnFilterHide}{sourceObsvnFilterHide}
}
\description{
Server function of the data - components
}
