% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{componentsUI}
\alias{componentsUI}
\title{Components UI}
\usage{
componentsUI(id, title = NULL)
}
\arguments{
\item{id}{id of module}

\item{title}{(character) title of module}
}
\description{
UI of the data - components tab
}
