% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{concentrationsServer}
\alias{concentrationsServer}
\title{Concentrations Server}
\usage{
concentrationsServer(id, config, values, events)
}
\arguments{
\item{id}{id of module}

\item{config}{(list) list of configuration parameters}

\item{values}{values}

\item{events}{events}
}
\description{
Server function of the data - concentrations
}
