% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dummy.R
\name{createDummyMatrix}
\alias{createDummyMatrix}
\title{Create dummy matrix}
\usage{
createDummyMatrix(names, current = NULL)
}
\arguments{
\item{names}{list with name vectors}

\item{current}{current matrix list}
}
\description{
Create dummy matrix
}
