% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-model.R
\name{createModel}
\alias{createModel}
\alias{alterModel}
\title{Create or Alter a Model Object}
\usage{
createModel(mode = "Run_update", notes = character(0), file = NULL)

alterModel(model, mode = NULL, notes = NULL)
}
\arguments{
\item{mode}{How should the model be run? One of Run_update, Run_independent,
Run_baseline}

\item{notes}{Notes to save with model}

\item{file}{Read model specifications from file}

\item{model}{Model Object}
}
\description{
Create or Alter a Model Object
}
\section{Functions}{
\itemize{
\item \code{alterModel}: 
}}

