% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-createModelCode.R
\name{createModelCode}
\alias{createModelCode}
\title{Create model code}
\usage{
createModelCode(
  priors,
  userEstimates,
  valueNames,
  constants,
  individualNames,
  modelOptions,
  covariates,
  termsSources,
  termsTargets,
  covariatesNum = NULL
)
}
\arguments{
\item{priors, }{character vector of priors from shiny interface}

\item{userEstimates, }{character vector of priors from shiny interface}

\item{valueNames, }{list of value names}

\item{constants, }{list of constants}

\item{individualNames, }{names of the individuals}

\item{modelOptions, }{list of model options}

\item{covariates, }{character matrix of categorical covariates}

\item{termsSources, }{list of sourceTerms}

\item{termsTargets, }{list of termsTargets}

\item{covariatesNum, }{numeric atrix of numeric covariates}
}
\value{
Nimble code to be used in \code{\link[nimble]{nimbleModel}}
}
\description{
Create model code
}
