% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-nameEvents.R
\name{createNameEvents}
\alias{createNameEvents}
\title{Create events from changed matrix names}
\usage{
createNameEvents(old, new, row, col, update = TRUE)
}
\arguments{
\item{old}{old matrix}

\item{new}{new matrix}

\item{row}{name of variable for row names to be used in events list}

\item{col}{name of variable for column names to be used in events list}

\item{update}{create update events}
}
\value{
list of events of the form list(event, variable, old, new)
}
\description{
Create events from changed matrix names
}
