% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-oxcal.R
\name{createOxCalText}
\alias{createOxCalText}
\title{Create Oxcal Text}
\usage{
createOxCalText(
  model,
  basicCode,
  terrestrialCurve,
  aquaticCurve1,
  aquaticCurve2,
  OxCalA,
  meanDeltaR1,
  sdDeltaR1,
  OxCalB,
  meanDeltaR2,
  sdDeltaR2,
  bins,
  coordinates
)
}
\arguments{
\item{model}{output of the model}

\item{basicCode}{(character) basic text read from external source}

\item{terrestrialCurve}{(character) basic text for the terrestrial curve read from external source}

\item{aquaticCurve1}{(character) basic text for the aquatic curve 1 read from external source}

\item{aquaticCurve2}{(character) basic text for the aquatic curve 2 read from external source}

\item{OxCalA}{(character) parameter estimate for aquatic curve 1}

\item{meanDeltaR1}{(numeric) input of the mean for aquatic curve 1}

\item{sdDeltaR1}{(numeric) input of the sd for aquatic curve 1}

\item{OxCalB}{(character) parameter estimate for aquatic curve 2}

\item{meanDeltaR2}{(numeric) input of the mean for aquatic curve 2}

\item{sdDeltaR2}{(numeric) input of the sd for aquatic curve 2}

\item{bins}{(character) either "meansd" for the usage of mean and sd, or "bins" for the usage
of pdf for the selected parameter estimate(s)}

\item{coordinates}{(data.frame) containing the radiocarbon values (mean+SD) for each target}
}
\description{
Create Oxcal Text
}
