% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-outputPlot.R
\name{extractContributionLimit}
\alias{extractContributionLimit}
\title{Extract contribution limit}
\usage{
extractContributionLimit(estType, userEstimateGroups)
}
\arguments{
\item{estType}{(character) type of estimates, e.g. "Source contributions", 
"Component contributions", ...}

\item{userEstimateGroups}{(list) list of user estimate groups}
}
\description{
Extract contribution limit
}
