% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-server.R
\name{extractPotentialCat}
\alias{extractPotentialCat}
\title{Extract Potential Cat}
\usage{
extractPotentialCat(targetValuesCovariates)
}
\arguments{
\item{targetValuesCovariates}{table with covariates}
}
\description{
Extract potential categorical covariates.
}
