% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-server.R
\name{extractPotentialNumerics}
\alias{extractPotentialNumerics}
\title{Extract Potential Numerics}
\usage{
extractPotentialNumerics(targetValuesCovariates)
}
\arguments{
\item{targetValuesCovariates}{table with covariates.}
}
\description{
Extract potential numerical covariates.
}
