% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fillFromUpload.R
\name{fillValuesFromUpload}
\alias{fillValuesFromUpload}
\title{Fill Values From Upload}
\usage{
fillValuesFromUpload(uploadedValues)
}
\arguments{
\item{uploadedValues}{list of uploaded values}
}
\value{
list of values
}
\description{
Fill Values From Upload
}
