% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-doubleMatrix.R
\name{fixMatrixCols}
\alias{fixMatrixCols}
\title{Fix Matrix Cols}
\usage{
fixMatrixCols(m, oldNames, fixedCols = FALSE, row, col)
}
\arguments{
\item{m}{matrix}

\item{oldNames}{(character) old column names, e.g. of "meanData()" or "covarianceData()"}

\item{fixedCols}{either FALSE or a character vector of column names to be kept, e.g. "Offset",
"longitude", "latitude", ...}

\item{row}{(character) row variable, e.g. "obsvnNames"}

\item{col}{(character) column variable, e.g. "targetNames"}
}
\description{
Fix columns of a matrix. If fixed colums are specified, they are used as column names otherwise
oldNames are kept.
}
