% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fruitsClass.R
\name{fruits}
\alias{fruits}
\title{Constructor for S3 class \code{fruits}}
\usage{
fruits(
  data,
  modelOptions,
  valueNames,
  priors = list(),
  userEstimates = list(list(), list())
)
}
\arguments{
\item{data}{list: input data}

\item{modelOptions}{list: contains elements \code{burnin} and \code{iterations}}

\item{valueNames}{list: contains names for the \code{targets},
\code{fractions}, and \code{sources}}

\item{priors}{list of characters with prior expressions}

\item{userEstimates}{list of characters}
}
\description{
All arguments are checked for correct class, length, names etc.
Constants and model code are added automatically.
}
\examples{
\dontrun{
# load fruits data:
load(system.file("app/exampleModels/", "fruitsExample.Rdata", package = "ReSources"))
# get Fruits-Object
fruitsObj <- fruits(data, modelOptions, valueNames, priors, userEstimates)
# run model
model <- compileRunModel(fruitsObj)
modelResults <- getResultStatistics(model$parameters, model$userEstimateSamples, fruitsObj,
  DT = FALSE, agg = FALSE
)

# better: use shiny app:
shiny::runApp(paste0(system.file(package = "ReSources"), "/app"))
}
}
