% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-server.R
\name{fruitsTab}
\alias{fruitsTab}
\title{Server Function for the FRUITS Module / Tab}
\usage{
fruitsTab(
  input,
  output,
  session,
  isoMemoData = function() {
     list(data = NULL, event = NULL)
 },
  isoDataExport = function() {
     list(data = NULL, event = NULL)
 }
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{isoMemoData}{data from IsoMemo App}

\item{isoDataExport}{data to export to IsoMemo App}
}
\description{
Server Function for the FRUITS Module / Tab
}
