% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ui.R
\name{fruitsUI}
\alias{fruitsUI}
\title{UI function for the FRUITS Tab / Module}
\usage{
fruitsUI(id, title = "FRUITS")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
UI function for the FRUITS Tab / Module
}
