% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-oxcal.R
\name{getCurveFormula}
\alias{getCurveFormula}
\title{Get Curve Formula}
\usage{
getCurveFormula(curve, mean = NULL, sd = NULL)
}
\arguments{
\item{curve}{(list of strings) parsed from one section of one of the curve files}

\item{mean}{(numeric) optional input of the mean for the curve}

\item{sd}{(numeric) optional input of the sd for the curve}
}
\description{
Get Curve Formula
}
