% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{getFractionNames}
\alias{getFractionNames}
\title{Get Fraction Names}
\usage{
getFractionNames(
  withComponents,
  withConcentration,
  weights,
  concentration,
  targetNames
)
}
\arguments{
\item{withComponents}{(logical) TRUE if "Include components"}

\item{withConcentration}{(logical) TRUE if "Include concentration"}

\item{weights}{(list) list of weights tables}

\item{concentration}{(list) list of concentration tables}

\item{targetNames}{(character) names of the targets}
}
\description{
Get Fraction Names
}
