% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-oxcal.R
\name{getLoop}
\alias{getLoop}
\title{Get Loop}
\usage{
getLoop(optionCurve1, optionCurve2, model, bins, OxCalA, OxCalB, coordinates)
}
\arguments{
\item{optionCurve1}{(character) option text for the aquatic curve 1 read from external source}

\item{optionCurve2}{(character) option text for the aquatic curve 2 read from external source}

\item{model}{output of the model}

\item{bins}{(character) either "meansd" for the usage of mean and sd, or "bins" for the usage
of pdf for the selected parameter estimate(s)}

\item{OxCalA}{(character) parameter estimate for aquatic curve 1}

\item{OxCalB}{(character) parameter estimate for aquatic curve 2}

\item{coordinates}{(data.frame) containing the radiocarbon values (mean+SD) for each target}
}
\description{
Loop over all targets
}
