% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-modelSummary.R
\name{getResultStatistics}
\alias{getResultStatistics}
\title{Analyse MCMC results}
\usage{
getResultStatistics(
  parameters,
  userEstimates,
  fruitsObj,
  statistics = NULL,
  DT = TRUE,
  agg = TRUE,
  bins = FALSE
)
}
\arguments{
\item{parameters}{parameters object of ReSources model object}

\item{userEstimates}{userEstimateSamples object of ReSources model object}

\item{fruitsObj}{object of class fruits: input data}

\item{statistics}{character vector only if agg = TRUE, compute aggregate statistics,
use R functions such as mean or sd}

\item{DT}{boolean export as data table}

\item{agg}{boolean aggregate data by statistcs}

\item{bins}{boolean quantile bins 0-100\%}
}
\description{
Get results statistics from ReSources  model object
}
