% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{getSourceNames}
\alias{getSourceNames}
\title{Get Source Names}
\usage{
getSourceNames(withConcentration, concentration, source)
}
\arguments{
\item{withConcentration}{(logical) TRUE if "Include concentrations"}

\item{concentration}{(list) list of concentration tables}

\item{source}{(list) list of source tables}
}
\description{
Get Source Names
}
