% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{isDeepestEntry}
\alias{isDeepestEntry}
\title{Is Deepest Entry}
\usage{
isDeepestEntry(entryContent)
}
\arguments{
\item{entryContent}{(list) element of values, e.g. values$source, values$sourceUncert,
values$sourceOffset, values$sourceOffsetUncert}
}
\description{
Checks if names of the list are targetNames (deepest hierarchy in a values object)
}
