% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-server.R
\name{isObsvnNames}
\alias{isObsvnNames}
\title{Is Obsvn Names}
\usage{
isObsvnNames(entryContent, lengthNewNames)
}
\arguments{
\item{entryContent}{(list) element of values, e.g. values$source, values$sourceUncert,
values$sourceOffset, values$sourceOffsetUncert}

\item{lengthNewNames}{(numeric) number of new names}
}
\description{
Checks if names if the list are obsvnNames (hierarchy above targetNames in a values object)
}
