% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{isPreDeepestEntry}
\alias{isPreDeepestEntry}
\title{Is Pre Deepest Entry}
\usage{
isPreDeepestEntry(entryContent)
}
\arguments{
\item{entryContent}{(list) element of values, e.g. values$source, values$sourceUncert,
values$sourceOffset, values$sourceOffsetUncert}
}
\description{
Checks if names if the list are obsvnNames (hierarchy above targetNames in a values object)
}
