% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-nameEvents.R
\name{processNameEvents}
\alias{processNameEvents}
\title{Update matrix based on events list from createNameEvents}
\usage{
processNameEvents(m, events, row, col)
}
\arguments{
\item{m}{matrix}

\item{events}{event list from createNameEvents}

\item{row}{variable used for row names}

\item{col}{variable used for column names}
}
\value{
altered matrix
}
\description{
Update matrix based on events list from createNameEvents
}
