% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-readString.R
\name{readStringWrapper}
\alias{readStringWrapper}
\title{Read String Wrapper}
\usage{
readStringWrapper(content, mode, class)
}
\arguments{
\item{content}{(character) string from clipboard}

\item{mode}{(character) paste mode, one of "auto", "comma-separated", "tab-separated", "semicolon"}

\item{class}{(character) class of content, e.g. "numeric", "character"}
}
\description{
Read and check string from clipboard
}
