% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-createModelCode.R
\name{readSystemFile}
\alias{readSystemFile}
\title{Read system file form ReSources package}
\usage{
readSystemFile(path)
}
\arguments{
\item{path}{character: path relative to inst directory}
}
\description{
Wrapper for \code{\link[base]{readLines}} and
\code{\link[base]{system.file}}
}
