% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-fruitsMatrix.R
\name{removeLine}
\alias{removeLine}
\title{Remove Line}
\usage{
removeLine(matrix, type, name)
}
\arguments{
\item{matrix}{matrix}

\item{type}{character type of the line, either "row" or "column"}

\item{name}{character name of row or column}
}
\description{
Removes the named row or column from matrix
}
