% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{removeObsvnFromEntry}
\alias{removeObsvnFromEntry}
\title{Remove Obsvn From Entry}
\usage{
removeObsvnFromEntry(values, entry, name)
}
\arguments{
\item{values}{(list) list containing all input data (all input tables)}

\item{entry}{(character) one of c("source", "sourceUncert", "sourceOffset", "sourceOffsetUncert",
"sourceCovariance", "concentration", "concentrationUncert", "concentrationCovariance")}

\item{name}{(character) name of the target to be removed}
}
\description{
Removes all list entries named "name" from the element "entry".
}
