% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{removeObsvnFromLists}
\alias{removeObsvnFromLists}
\title{Remove Obsvn From Lists}
\usage{
removeObsvnFromLists(values, name)
}
\arguments{
\item{values}{(list) list containing all input data (all input tables)}

\item{name}{(character) name of the target to be removed}
}
\description{
Removes all list entries named "name" from all lists that contain entries named by observations.
}
