% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{resources.options}
\alias{resources.options}
\alias{resources.debug}
\alias{R_RESOURCES_DEBUG}
\title{Options used by the 'resources' package}
\description{
Below are all \R options specific to the \pkg{resources} package.
\cr
_WARNING: Note that the names and the default values of
these options may change in future versions of the package.
Please use with care until further notice._
}
\details{
\describe{
 \item{\option{resources.debug}:}{If `TRUE`, extensive debug messages
       are generated. (Default: `FALSE`)}
}
}
\keyword{internal}
