% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-shinyInputToClass.R
\name{shinyInputToClass}
\alias{shinyInputToClass}
\title{Build object of class \code{\link{fruits}} from shiny input}
\usage{
shinyInputToClass(values, priors, userEstimates)
}
\arguments{
\item{values}{list: Shiny input}

\item{priors}{list: input$priors}

\item{userEstimates}{list: input$userEstimates}
}
\description{
Build object of class \code{\link{fruits}} from shiny input
}
