% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{sourcesServer}
\alias{sourcesServer}
\title{Sources Server}
\usage{
sourcesServer(
  id,
  values,
  events,
  hideTargetFilter,
  sourceCovNames,
  termChoices,
  sourceObsvnFilterChoices,
  sourceObsvnFilterHide
)
}
\arguments{
\item{id}{id of module}

\item{values}{values}

\item{events}{events}

\item{hideTargetFilter}{hideTargetFilter}

\item{sourceCovNames}{sourceCovNames}

\item{termChoices}{termChoices}

\item{sourceObsvnFilterChoices}{sourceObsvnFilterChoices}

\item{sourceObsvnFilterHide}{sourceObsvnFilterHide}
}
\description{
Server function of the data - sources
}
