% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-startApplication.R
\name{startApplication}
\alias{startApplication}
\title{Start Application}
\usage{
startApplication(
  port = 4242,
  host = "127.0.0.1",
  launch.browser = getOption("shiny.launch.browser", interactive())
)
}
\arguments{
\item{port}{port of web application}

\item{host}{accept connections from this address}

\item{launch.browser}{whether to launch the application in a web browser}
}
\description{
Start Application
}
