% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{targetValuesUI}
\alias{targetValuesUI}
\title{Target Values UI}
\usage{
targetValuesUI(id, title = NULL)
}
\arguments{
\item{id}{id of module}

\item{title}{(character) title of module}
}
\description{
UI of the data - target values
}
