% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-server.R
\name{updateListNames}
\alias{updateListNames}
\title{Update List Names}
\usage{
updateListNames(entryContent, depth, newNames)
}
\arguments{
\item{entryContent}{(list) possibly nested list}

\item{depth}{depth of list where names should be updated}

\item{newNames}{new names for list elements}
}
\description{
Update List Names
}
