% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-server.R
\name{updateNamesIfMismatch}
\alias{updateNamesIfMismatch}
\title{Update Names If Mismatch}
\usage{
updateNamesIfMismatch(values, entryName)
}
\arguments{
\item{values}{(list) with all data and model options}

\item{entryName}{(character) name of values element}
}
\description{
Update term names of entries such that they match to new targetNames
}
