% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{updateObsvnsInLists}
\alias{updateObsvnsInLists}
\title{Remove Obsvn From Lists}
\usage{
updateObsvnsInLists(values, name, updateFun)
}
\arguments{
\item{values}{(list) list containing all input data (all input tables)}

\item{name}{(character) name of the target to be removed}

\item{updateFun}{name of function that updates the lists, either
deleteTableFromList or updateListNames}
}
\description{
Removes all list entries named "name" from all lists that contain entries named by observations.
}
