% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{updateObsvnsOfEntry}
\alias{updateObsvnsOfEntry}
\title{Remove Obsvn From Entry}
\usage{
updateObsvnsOfEntry(values, entry, name, updateFun)
}
\arguments{
\item{values}{(list) list containing all input data (all input tables)}

\item{entry}{(character) one of c("source", "sourceUncert", "sourceOffset", "sourceOffsetUncert",
"sourceCovariance", "concentration", "concentrationUncert", "concentrationCovariance")}

\item{name}{(character) name of the target to be removed}

\item{updateFun}{name of function that updates the lists, either
deleteTableFromList or updateListNames}
}
\description{
Removes all list entries named "name" from the element "entry".
}
