% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-server.R
\name{updateSourceNamesIfMismatch}
\alias{updateSourceNamesIfMismatch}
\title{Update Source Names If Mismatch}
\usage{
updateSourceNamesIfMismatch(values)
}
\arguments{
\item{values}{(list) with all data and model options}
}
\description{
Update term names of source related entries such that they match to new targetNames
}
