% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{updateTargetsInLists}
\alias{updateTargetsInLists}
\title{Update Targets In Lists}
\usage{
updateTargetsInLists(values, name, updateFun)
}
\arguments{
\item{values}{(list) list containing all input data (all input tables)}

\item{name}{(character) name of the target to be removed (only )}

\item{updateFun}{name of function that updates the lists, either
deleteTableFromList or updateListNames}
}
\description{
Update all list entries named "name" from all lists that contain entries named by targets.
}
