% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-server.R
\name{areNamesNotMatching}
\alias{areNamesNotMatching}
\title{Update Names If Mismatch}
\usage{
areNamesNotMatching(values, entryName, newNames, isNamesFun)
}
\arguments{
\item{values}{(list) with all data and model options}

\item{entryName}{(character) name of values element}

\item{newNames}{(reactive) character vector of newNames}

\item{isNamesFun}{(function) function that checks for the correct level in the list hierarchy
of values}
}
\description{
Update names of entries such that they match to new newNames
}
