% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{concentrationsServer}
\alias{concentrationsServer}
\title{Concentrations Server}
\usage{
concentrationsServer(id, values, events, hideTargetFilter, baselineModel)
}
\arguments{
\item{id}{id of module}

\item{values}{values}

\item{events}{events}

\item{hideTargetFilter}{hideTargetFilter}

\item{baselineModel}{TRUE if baselineModel}
}
\description{
Server function of the data - concentrations
}
