% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{concentrationsServer}
\alias{concentrationsServer}
\title{Concentrations Server}
\usage{
concentrationsServer(
  id,
  values,
  events,
  hideTargetFilter,
  sourceObsvnFilterChoices,
  sourceObsvnFilterHide
)
}
\arguments{
\item{id}{id of module}
}
\description{
Server function of the data - concentrations
}
