% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{deleteTableFromList}
\alias{deleteTableFromList}
\title{Delete Table From List}
\usage{
deleteTableFromList(entryContent, depth, name)
}
\arguments{
\item{entryContent}{(list) possibly nested list}

\item{depth}{depth of list where names should be updated}

\item{name}{name of list element to be deleted}
}
\description{
Delete Table From List
}
