% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{getDepthAndTable}
\alias{getDepthAndTable}
\title{Get Depth And Table}
\usage{
getDepthAndTable(entryContent, isEntryFun = isDeepestEntry, n = NULL)
}
\arguments{
\item{entryContent}{(list) list to look for names}

\item{isEntryFun}{(function) function that checks for the correct level in the list hierarchy}

\item{n}{(numeric) depth of list to look for names
of values}
}
\description{
Get the depth of the nested list and the content of the most inner table.
}
