% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{sourcesServer}
\alias{sourcesServer}
\title{Sources Server}
\usage{
sourcesServer(id, config, values, events, termChoices)
}
\arguments{
\item{id}{id of module}

\item{config}{(list) list of configuration parameters}

\item{values}{values}

\item{events}{events}

\item{termChoices}{termChoices}
}
\description{
Server function of the data - sources
}
