% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{sourcesServer}
\alias{sourcesServer}
\title{Sources Server}
\usage{
sourcesServer(
  id,
  values,
  events,
  withComponents,
  hideTargetFilter,
  termChoices,
  sourceObsvnFilterChoices,
  sourceObsvnFilterHide
)
}
\arguments{
\item{id}{id of module}

\item{values}{values}

\item{events}{events}

\item{withComponents}{(reactive) logical, TRUE if include components}

\item{hideTargetFilter}{(reactive) logical, hideTargetFilter}

\item{termChoices}{termChoices}

\item{sourceObsvnFilterChoices}{sourceObsvnFilterChoices}

\item{sourceObsvnFilterHide}{sourceObsvnFilterHide}
}
\description{
Server function of the data - sources
}
