% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{sourcesServer}
\alias{sourcesServer}
\title{Sources Server}
\usage{
sourcesServer(id, values, events, hideTargetFilter, termChoices, baselineModel)
}
\arguments{
\item{id}{id of module}

\item{values}{values}

\item{events}{events}

\item{hideTargetFilter}{(reactive) logical, hideTargetFilter}

\item{termChoices}{termChoices}

\item{baselineModel}{TRUE if baselineModel}
}
\description{
Server function of the data - sources
}
