% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-startApplication.R
\name{startApplication}
\alias{startApplication}
\title{Start Application}
\usage{
startApplication(port = 4242, host = "127.0.0.1")
}
\arguments{
\item{port}{port of web application}

\item{host}{accept connections from this address}
}
\description{
Start Application
}
