% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-data.R
\name{targetValuesServer}
\alias{targetValuesServer}
\title{Target Values Server}
\usage{
targetValuesServer(id, values, events, termChoices, modelType)
}
\arguments{
\item{id}{id of module}

\item{values}{values}

\item{events}{events}

\item{termChoices}{termChoices}
}
\description{
Server function of the data - target values
}
