% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-updateListNames.R
\name{updateListNames}
\alias{updateListNames}
\title{Update List Names}
\usage{
updateListNames(entryContent, depth, namesList)
}
\arguments{
\item{entryContent}{(list) possibly nested list}

\item{depth}{depth of list where names should be updated}

\item{namesList}{new names for list elements}
}
\description{
Update List Names
}
