% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{uploadModelUI}
\alias{uploadModelUI}
\alias{uploadModel}
\title{Upload model module}
\usage{
uploadModelUI(id)

uploadModel(input, output, session, values, model, uploadedNotes, reset)
}
\arguments{
\item{id}{id of module}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{values}{(reactive) list: Shiny input}

\item{model}{(reactive) output of the model}

\item{uploadedNotes}{(reactive) variable that stores content of README.txt}

\item{reset}{(reactive) result of reset button}
}
\description{
UI function to upload a zip file with notes and a list of models

Backend for upload model module
}
